+++ pico-rv32ima/emulator.h
@@ -3,8 +3,6 @@
 
 #include <stdint.h>
 
-#define RAM_MB 16
-
 extern uint32_t ram_amt;
 void rvEmulator();
